requirements.txt
aiocache==0.12.3
aiocqhttp==1.4.4
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiolimiter==1.1.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
arclet-alconna==1.8.32
arclet-alconna-tools==0.7.10
argcomplete==3.5.1
arrow==1.3.0
async-timeout==4.0.3
atlastk==0.13.3
attrs==24.2.0
bce-python-sdk==0.9.23
binaryornot==0.4.4
blinker==1.8.2
Brotli==1.1.0
cachetools==5.5.0
cashews==7.3.2
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
cryptography==43.0.3
dill==0.3.9
diskcache==5.6.3
distlib==0.3.9
exceptiongroup==1.2.2
fastapi==0.115.4
filelock==3.16.1
filetype==1.2.0
Flask==3.0.3
frozenlist==1.5.0
future==1.0.0
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
inexactsearch==1.0.2
itsdangerous==2.2.0
Jinja2==3.1.4
linkify-it-py==2.0.3
loguru==0.7.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.17
nb-cli==1.4.2
nb-cli-plugin-docker==0.4.1
nepattern==0.7.6
nonebot-adapter-console==0.6.0
nonebot-adapter-onebot==2.4.6
nonebot-plugin-alconna==0.53.1
nonebot-plugin-htmlrender==0.4.0
nonebot-plugin-send-anything-anywhere==0.7.1
nonebot_plugin_session==0.3.2
nonebot-plugin-waiter==0.8.0
nonebot2==2.4.0
nonechat==0.3.0
noneprompt==0.1.9
packaging==24.1
pillow==11.0.0
pip==24.2
pipx==1.7.1
platformdirs==4.3.6
playwright==1.48.0
priority==2.0.0
prompt_toolkit==3.0.48
propcache==0.2.0
py-cpuinfo==9.0.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
PyDeepLX==1.0.7
pyee==12.0.0
pyfiglet==1.0.2
Pygments==2.18.0
pygtrie==2.5.0
pymdown-extensions==10.12
pyOpenSSL==24.2.1
pyspellchecker==0.8.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-markdown-math==0.8
python-slugify==8.0.4
PyYAML==6.0.2
qianfan==0.4.11
Quart==0.19.8
requests==2.32.3
rich==13.9.4
setuptools==75.1.0
shellingham==1.5.4
silpa_common==0.3
six==1.16.0
sniffio==1.3.1
socksio==1.0.0
soundex==1.1.3
starlette==0.41.2
StrEnum==0.4.15
tarina==0.6.3
taskgroup==0.0.0a4
tenacity==8.5.0
text-unidecode==1.3
textual==0.85.2
tomli==2.0.2
tomlkit==0.13.2
tortoise==0.1.1
typer==0.12.5
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
uc-micro-py==1.0.3
urllib3==2.2.3
userpath==1.9.2
uvicorn==0.32.0
virtualenv==20.27.1
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
Werkzeug==3.1.1
wheel==0.44.0
win32-setctime==1.1.0
wsproto==1.2.0
yarl==1.17.1
zipp==3.20.2
